Add getNodeList to configserverclient
