+++ controller-api/src/main/java/com/yahoo/vespa/hosted/controller/api/integration/configserver/ConfigServerClient.java
@@ -67,5 +67,12 @@ public interface ConfigServerClient {
      */
     EndpointStatus getGlobalRotationStatus(DeploymentId deployment, String endpoint) throws IOException;
 
-    //JsonNode getClusterInfo(DeploymentId deployment, String endpoint) throws IOException;
+    /**
+     * The nodes allocated to the deployment
+     *
+     * @param deployment The application/zone pair
+     * @return The nodes for this deployment
+     * @throws IOException If unable to retrieve the node list
+     */
+    NodeList getNodeList(DeploymentId deployment) throws IOException;
 }
