+++ controller-server/src/main/java/com/yahoo/vespa/hosted/controller/maintenance/ClusterInfoMaintainer.java
@@ -1,24 +1,21 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.vespa.hosted.controller.maintenance;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.yahoo.config.provision.ApplicationId;
 import com.yahoo.config.provision.ClusterSpec;
-import com.yahoo.config.provision.Zone;
 import com.yahoo.vespa.curator.Lock;
 import com.yahoo.vespa.hosted.controller.Application;
 import com.yahoo.vespa.hosted.controller.Controller;
+import com.yahoo.vespa.hosted.controller.api.identifiers.DeploymentId;
+import com.yahoo.vespa.hosted.controller.api.integration.configserver.NodeList;
 import com.yahoo.vespa.hosted.controller.application.ClusterInfo;
 import com.yahoo.vespa.hosted.controller.application.Deployment;
 
 import java.io.IOException;
-import java.net.URI;
 import java.time.Duration;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
 /**
@@ -37,24 +34,24 @@ public class ClusterInfoMaintainer extends Maintainer {
         this.controller = controller;
     }
 
-    private static String clusterid(NodeRepositoryJsonModel.Node node) {
+    private static String clusterid(NodeList.Node node) {
         return node.membership.clusterId;
     }
 
-    private Map<ClusterSpec.Id, ClusterInfo> getClusterInfo(NodeRepositoryJsonModel nodes) {
+    private Map<ClusterSpec.Id, ClusterInfo> getClusterInfo(NodeList nodes) {
         Map<ClusterSpec.Id, ClusterInfo> infoMap = new HashMap<>();
 
         // Group nodes by clusterid
-        Map<String, List<NodeRepositoryJsonModel.Node>> clusters = nodes.nodes.stream()
+        Map<String, List<NodeList.Node>> clusters = nodes.nodes.stream()
                 .filter(node -> node.membership != null)
                 .collect(Collectors.groupingBy(ClusterInfoMaintainer::clusterid));
 
         // For each cluster - get info
         for (String id : clusters.keySet()) {
-            List<NodeRepositoryJsonModel.Node> clusterNodes = clusters.get(id);
+            List<NodeList.Node> clusterNodes = clusters.get(id);
 
             //Assume they are all equal and use first node as a representatitve for the cluster
-            NodeRepositoryJsonModel.Node node = clusterNodes.get(0);
+            NodeList.Node node = clusterNodes.get(0);
 
             // Add to map
             List<String> hostnames = clusterNodes.stream().map(node1 -> node1.hostname).collect(Collectors.toList());
@@ -65,58 +62,21 @@ public class ClusterInfoMaintainer extends Maintainer {
         return infoMap;
     }
 
-    // TODO use appId in url
-    private NodeRepositoryJsonModel getApplicationNodes(ApplicationId appId, Zone zone) {
-        NodeRepositoryJsonModel nodesResponse = null;
-        ObjectMapper mapper = new ObjectMapper();
-        for (URI uri : controller.getConfigServerUris(zone.environment(), zone.region())) {
-            try {
-                nodesResponse = mapper.readValue(uri.toURL(), NodeRepositoryJsonModel.class);
-                break;
-            } catch (IOException ioe) {
-                //TODO
-            }
-        }
-        return nodesResponse;
-    }
-
     @Override
     protected void maintain() {
 
         for (Application application : controller().applications().asList()) {
             Lock lock = controller().applications().lock(application.id());
             for (Deployment deployment : application.deployments().values()) {
-                NodeRepositoryJsonModel appNodes = getApplicationNodes(application.id(), deployment.zone());
-                Map<ClusterSpec.Id, ClusterInfo> clusterInfo = getClusterInfo(appNodes);
-                Application app = application.with(deployment.withClusterInfo(clusterInfo));
-                controller.applications().store(app, lock);
-            }
-        }
-    }
-
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    public static class NodeRepositoryJsonModel {
-
-        @JsonProperty("nodes")
-        public List<Node> nodes;
-
-        @JsonIgnoreProperties(ignoreUnknown = true)
-        public static class Node {
-            @JsonProperty("hostname")
-            public String hostname;
-            @JsonProperty("flavor")
-            public String flavor;
-            @JsonProperty("membership")
-            public Membership membership;
-            @JsonProperty("cost")
-            public int cost;
-
-            @JsonIgnoreProperties(ignoreUnknown = true)
-            public static class Membership {
-                @JsonProperty("clustertype")
-                public String clusterType;
-                @JsonProperty("clusterid")
-                public String clusterId;
+                DeploymentId deploymentId = new DeploymentId(application.id(), deployment.zone());
+                try {
+                    NodeList nodes = controller().applications().configserverClient().getNodeList(deploymentId);
+                    Map<ClusterSpec.Id, ClusterInfo> clusterInfo = getClusterInfo(nodes);
+                    Application app = application.with(deployment.withClusterInfo(clusterInfo));
+                    controller.applications().store(app, lock);
+                } catch (IOException ioe) {
+                    Logger.getLogger(ClusterInfoMaintainer.class.getName()).fine(ioe.getMessage());
+                }
             }
         }
     }
